+++ cmd/explorer/src/connector/__init__.py
@@ -22,7 +22,6 @@ from dataclasses import asdict
 from typing import Any
 
 from aiohttp import ClientSession
-
 from cache import AsyncTTL
 
 path = os.path.join(os.path.dirname(__file__), "../../../../")
